update with personal keybindings for escape sequence
Update for jupyterlab3 - switch to pip install (federated) (#19)

* update for jupyterlab 3

following https://jupyterlab.readthedocs.io/en/latest/developer/extension_migration_2_3.html

By making the extension pip installable (federated) the way to access codemirror is a bit different as the extension is no longer included in the jupyterlab build step. So we use the `ensureVimKeymap` function on the ICodeMirror interface. This ensures that the codemirror used by jupyterlab has loaded it's vim bindings. We then access that codemirror via the ICodeMirror interface.

* appease eslint and prettier

Set eslint and prettier to the defaults in the extension cookiecutter during the upgrade, this applies them to index.ts.

* Update History.md

* Update node version in build.yml

* Fix linter script for PR action

* Add action to publish package to PyPi

* Update configurations and actions

Co-authored-by: Axel Fahy <axel@fahy.net>